+++ Sources/Common/Network/Connection/Connections/MistyswapConnection.swift
@@ -71,7 +71,9 @@ final class MistyswapConnection: Connection<
 
     func initiateOfframp(
         request: MistyswapOfframp_InitiateOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_InitiateOfframpResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_InitiateOfframpResponse, ConnectionError>
+        ) -> Void
     ) {
         guard config.mistyswapConfig() != nil else {
             completion(
@@ -96,7 +98,9 @@ final class MistyswapConnection: Connection<
 
     func getOfframpStatus(
         request: MistyswapOfframp_GetOfframpStatusRequest,
-        completion: @escaping (Result<MistyswapOfframp_GetOfframpStatusResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_GetOfframpStatusResponse, ConnectionError>
+        ) -> Void
     ) {
         guard config.mistyswapConfig() != nil else {
             completion(
@@ -121,7 +125,9 @@ final class MistyswapConnection: Connection<
 
     func forgetOfframp(
         request: MistyswapOfframp_ForgetOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>
+        ) -> Void
     ) {
         guard config.mistyswapConfig() != nil else {
             completion(
