+++ Sources/Common/Network/Service/FogServices.swift
@@ -72,27 +72,3 @@ protocol FogReportService {
         request: Report_ReportRequest,
         completion: @escaping (Result<Report_ReportResponse, ConnectionError>) -> Void)
 }
-
-protocol MistyswapService {
-    func initiateOfframp(
-        request: MistyswapOfframp_InitiateOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_InitiateOfframpResponse, ConnectionError>) -> Void)
-
-    func getOfframpStatus(
-        request: MistyswapOfframp_GetOfframpStatusRequest,
-        completion: @escaping (Result<MistyswapOfframp_GetOfframpStatusResponse, ConnectionError>) -> Void)
-
-    func forgetOfframp(
-        request: MistyswapOfframp_ForgetOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>) -> Void)
-}
-
-protocol MistyswapServiceConnection: MistyswapService, ConnectionProtocol {}
-
-// protocol MistyswapUntrustedService {
-//    func forgetOfframp(
-//        request: MistyswapOfframp_ForgetOfframpRequest,
-//        completion: @escaping (Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>) -> Void)
-// }
-//
-// protocol MistyswapUntrustedServiceConnection: MistyswapUntrustedService, ConnectionProtocol {}
