+++ Sources/GRPC/GRPC/GrpcConnection/GrpcConnections/EmptyMistyswapGrpcConnection.swift
@@ -31,7 +31,9 @@ final class EmptyMistyswapGrpcConnection: AttestedGrpcConnection, MistyswapServi
 
     func initiateOfframp(
         request: MistyswapOfframp_InitiateOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_InitiateOfframpResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_InitiateOfframpResponse, ConnectionError>
+        ) -> Void
     ) {
         completion(.failure(.connectionFailure(
             "Config used to intialize your client " +
@@ -41,7 +43,9 @@ final class EmptyMistyswapGrpcConnection: AttestedGrpcConnection, MistyswapServi
 
     func getOfframpStatus(
         request: MistyswapOfframp_GetOfframpStatusRequest,
-        completion: @escaping (Result<MistyswapOfframp_GetOfframpStatusResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_GetOfframpStatusResponse, ConnectionError>
+        ) -> Void
     ) {
         completion(.failure(.connectionFailure(
             "Config used to intialize your client " +
@@ -51,7 +55,9 @@ final class EmptyMistyswapGrpcConnection: AttestedGrpcConnection, MistyswapServi
 
     func forgetOfframp(
         request: MistyswapOfframp_ForgetOfframpRequest,
-        completion: @escaping (Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>) -> Void
+        completion: @escaping (
+            Result<MistyswapOfframp_ForgetOfframpResponse, ConnectionError>
+        ) -> Void
     ) {
         completion(.failure(.connectionFailure(
             "Config used to intialize your client " +
